[Chore] #22 실시간베스트아이템 더미 생성
