+++ CDS-APP-2-iOS/Presentation/HatCategory/Dummy/HeaderCategoryDummy.swift
@@ -26,18 +26,4 @@ extension HeaderCategory {
     }
 }
 
-struct RealtimeBestItem {
-    let image: UIImage
-    let brandName: String
-    let salePercent: String
-    let itemPrice: String
-    let rankingNumber: Int
-}
 
-extension RealtimeBestItem{
-    static func realtimeBestDummy() -> [RealtimeBestItem] {
-        return [
-
-        ]
-    }
-}
